Refactored ssl
